% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCluster.R
\name{calcCluster}
\alias{calcCluster}
\title{Calculating K-Means Clusters}
\usage{
calcCluster(data, clusters)
}
\arguments{
\item{data}{A dataframe with numbers.}

\item{clusters}{The number of clusters.}
}
\value{
A calculation of clusters.
}
\description{
Calculating K-Means Clusters
}
