% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCluster.R
\name{plotClusters}
\alias{plotClusters}
\title{Plot K-Means Clustering Results}
\usage{
plotClusters(data, clusters)
}
\arguments{
\item{data}{A dataframe with numbers.}

\item{clusters}{A k-means clustering object.}
}
\value{
A scatter plot with clusters.
}
\description{
Plot K-Means Clustering Results
}
